refactor(hd256/sm100): extract _as_bshkrd_tensor/_as_shhb_tensor into shared layout_utils module
Both helpers were duplicated across the forward and backward files after
the forward kernel was introduced. Move them to sm100_hd256_layout_utils.py
and import from there, eliminating ~110 lines of duplicated code.

Assertion messages ("varlen forward"/"varlen backward") are unified to the
neutral "varlen" since the functions are now shared. No kernel logic change.
