+++ flash_attn/cute/sm100_hd256_2cta_fmha_forward.py
@@ -31,46 +31,7 @@ from flash_attn.cute.tile_scheduler import SM100_TMEM_CAPACITY_COLUMNS
 from flash_attn.cute.flash_fwd_sm100 import DescaleTensors, _TUNING_CONFIG
 from flash_attn.cute.utils import ex2_emulation_2
 from flash_attn.cute.utils import AuxData
-
-
-def _as_bshkrd_tensor(
-    tensor: cute.Tensor,
-    h_k: Int32,
-    h_r: Int32,
-    varlen: bool,
-) -> cute.Tensor:
-    """Normalize (B,S,H,D)/(S,H,D) tensors to (B,S,H_k,H_r,D) preserving real strides."""
-    if cutlass.const_expr(cute.rank(tensor.layout) == 5):
-        return tensor
-    if cutlass.const_expr(cute.rank(tensor.layout) == 4):
-        return cute.make_tensor(
-            tensor.iterator,
-            cute.make_layout(
-                (tensor.shape[0], tensor.shape[1], h_k, h_r, tensor.shape[3]),
-                stride=(
-                    tensor.stride[0],
-                    tensor.stride[1],
-                    tensor.stride[2] * h_r,
-                    tensor.stride[2],
-                    tensor.stride[3],
-                ),
-            ),
-        )
-    assert cutlass.const_expr(cute.rank(tensor.layout) == 3), "Expected rank-3 varlen tensor"
-    assert cutlass.const_expr(varlen), "Rank-3 input is only valid for varlen forward"
-    return cute.make_tensor(
-        tensor.iterator,
-        cute.make_layout(
-            (1, tensor.shape[0], h_k, h_r, tensor.shape[2]),
-            stride=(
-                0,
-                tensor.stride[0],
-                tensor.stride[1] * h_r,
-                tensor.stride[1],
-                tensor.stride[2],
-            ),
-        ),
-    )
+from flash_attn.cute.sm100_hd256_layout_utils import _as_bshkrd_tensor
 
 
 class BlackwellFusedMultiHeadAttentionForward:
